# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aqary Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface PropertyItem {
  id: number;
  title: string;
  description: string;
  price: number;
  location: string;
  area: number;
  /** @nullable */
  rooms?: number | null;
  propertyType: string;
  features: string[];
  /** @nullable */
  imageUrl?: string | null;
  imageUrls?: string[];
  sellerId: number;
  status: string;
  views: number;
  saves: number;
  contacts: number;
  createdAt: string;
}

export type CreatePropertyBodyPropertyType =
  (typeof CreatePropertyBodyPropertyType)[keyof typeof CreatePropertyBodyPropertyType];

export const CreatePropertyBodyPropertyType = {
  apartment: "apartment",
  villa: "villa",
  commercial: "commercial",
  land: "land",
} as const;

export interface CreatePropertyBody {
  title: string;
  description: string;
  price: number;
  location: string;
  area: number;
  rooms?: number;
  propertyType: CreatePropertyBodyPropertyType;
  features?: string[];
  imageUrl?: string;
  /** @maxItems 3 */
  imageUrls?: string[];
}

export interface UpdatePropertyBody {
  title?: string;
  description?: string;
  price?: number;
  location?: string;
  area?: number;
  rooms?: number;
  propertyType?: string;
  features?: string[];
  imageUrl?: string;
  /** @maxItems 3 */
  imageUrls?: string[];
}

export interface RecommendedProperty {
  id: number;
  title: string;
  description: string;
  price: number;
  location: string;
  area: number;
  /** @nullable */
  rooms?: number | null;
  propertyType: string;
  features: string[];
  /** @nullable */
  imageUrl?: string | null;
  sellerId: number;
  status: string;
  views: number;
  saves: number;
  contacts: number;
  createdAt: string;
  matchScore: number;
  matchReasons: string[];
}

export type TrackInteractionBodyInteractionType =
  (typeof TrackInteractionBodyInteractionType)[keyof typeof TrackInteractionBodyInteractionType];

export const TrackInteractionBodyInteractionType = {
  view: "view",
  save: "save",
  unsave: "unsave",
  contact: "contact",
  scroll: "scroll",
  time_spent: "time_spent",
} as const;

export interface TrackInteractionBody {
  propertyId: number;
  interactionType: TrackInteractionBodyInteractionType;
  seconds?: number;
}

export type ChatHistoryItemRole =
  (typeof ChatHistoryItemRole)[keyof typeof ChatHistoryItemRole];

export const ChatHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatHistoryItem {
  role: ChatHistoryItemRole;
  content: string;
}

export interface ChatMessageBody {
  message: string;
  conversationHistory?: ChatHistoryItem[];
}

export interface ChatMatchedProperty {
  id: number;
  title: string;
  price: number;
  location: string;
  propertyType: string;
  /** Relative frontend URL for property details */
  propertyUrl: string;
  matchReasons: string[];
}

export interface ChatResponse {
  reply: string;
  properties?: ChatMatchedProperty[];
  feedbackCreated?: boolean;
}

export interface FeedbackItem {
  id: number;
  /** @nullable */
  userId?: number | null;
  message: string;
  /** @nullable */
  criteria?: string | null;
  resolved: boolean;
  createdAt: string;
}

export interface CreateFeedbackBody {
  message: string;
  criteria?: string;
}

export interface TrackPageViewBody {
  /** Client route path starting with / */
  path: string;
}

export interface PagePathCount {
  path: string;
  count: number;
}

export interface PropertyLeaderItem {
  id: number;
  title: string;
  views: number;
  saves: number;
  contacts: number;
  sellerId: number;
}

export interface UserActivityItem {
  userId: number;
  name: string;
  email: string;
  interactionCount: number;
  pageViewCount: number;
}

export interface SellerActivityEvent {
  id: number;
  propertyId: number;
  propertyTitle: string;
  interactionType: string;
  createdAt: string;
}

export interface SellerActivityResponse {
  recentInteractions: SellerActivityEvent[];
}

export type AnalyticsDataInteractionsByType = { [key: string]: number };

export interface AnalyticsData {
  totalUsers: number;
  totalProperties: number;
  totalInteractions: number;
  pendingProperties: number;
  unresolvedFeedbacks: number;
  totalPageViews: number;
  interactionsByType: AnalyticsDataInteractionsByType;
  pageViewsByPath: PagePathCount[];
  topPropertiesByViews: PropertyLeaderItem[];
  userActivity: UserActivityItem[];
}

export interface UserPreferencesData {
  /** @nullable */
  preferredLocation?: string | null;
  /** @nullable */
  maxBudget?: number | null;
  /** @nullable */
  preferredType?: string | null;
  preferredFeatures: string[];
}

export interface UpdatePreferencesBody {
  preferredLocation?: string;
  maxBudget?: number;
  preferredType?: string;
  preferredFeatures?: string[];
}

export type ListPropertiesParams = {
  search?: string;
  type?: string;
  location?: string;
  minPrice?: number;
  maxPrice?: number;
};