# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aqary Real Estate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /track:
    post:
      operationId: trackInteraction
      tags: [interactions]
      summary: Track user interaction with property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackInteractionBody"
      responses:
        "201":
          description: Interaction tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /track/page-view:
    post:
      operationId: trackPageView
      tags: [interactions]
      summary: Track SPA page visit (optional auth for user attribution)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackPageViewBody"
      responses:
        "201":
          description: Page view recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        role:
          type: string
          enum: [buyer, seller]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    TrackInteractionBody:
      type: object
      properties:
        propertyId:
          type: integer
        interactionType:
          type: string
          enum: [view, save, unsave, contact, scroll, time_spent]
        seconds:
          type: number
      required:
        - propertyId
        - interactionType
...
    TrackPageViewBody:
      type: object
      properties:
        path:
          type: string
          description: Client route path starting with /
      required:
        - path
...